requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
psutil
requests
aiohttp
playwright
watchdog
cryptography
apscheduler
numpy
pandas
openai
pyttsx3
SpeechRecognition
pyaudio
pyobjc-framework-Quartz
pyobjc-framework-Vision
pyobjc-framework-Cocoa
faiss-cpu
python-dotenv
beautifulsoup4
pillow
qrcode
netifaces
scapy
playwright
